Add files via upload
In now days all are somehow stressed  and causing health issues and this is going worsen day by day. All are busy on that extent level that they can’t even take care of their health and life is somehow getting effected by this.There have been many health tracker available that have been developed and monitoring health of a person and in case of unlike detects in body warning is displayed. But through processing and  analyzing by using ML and DLA we are making watch as self-predicting device which can predict own their own.We are using A.I in order to built a watch which monitors health of an individual , this watch basically collect data such as heart rate , sleep cycle, breathing rate ,blood pressure , activity cycle , oxygen rate etc using various sensors.